feat: Successfully Parking Class, Service and Controller

+++ parking/src/main/java/br/com/escuderodev/parking/models/parking/ParkingListData.java
@@ -11,7 +11,7 @@ public record ParkingListData(
         BigDecimal variableParkingPrice,
         LocalDateTime startParking,
         LocalDateTime stopParking,
-        Integer usageTime,
+        Long usageTime,
         BigDecimal amountToPay,
         Vehicle vehicle
 ) {
